update action modal fixed

+++ bot/views.py
@@ -72,8 +72,6 @@ def register(request):
         password = request.POST.get('password')
         confirm_password = request.POST.get('confirm_password')
 
-        print(username, email, password, confirm_password)
-
         if password == confirm_password:
             try:
                 new_user = Users(
@@ -97,28 +95,33 @@ def signout(request):
 
 # function for editing data
 def edit_customer(request, user_id: int):
-    if request.method == 'POST':
-        print(user_id)
-        customer = get_object_or_404(Customer, user_id=user_id)
-        data = json.loads(request.body)
+    if "username" in request.session:
+        if request.method == 'POST':
+            customer = get_object_or_404(Customer, user_id=user_id)
+            data = json.loads(request.body)
 
-        # Update the customer with new data
-        customer.person_name = data.get('person_name')
-        customer.phone_number = data.get('phone_number')
-        customer.age = data.get('age')
-        customer.appointment_date = data.get('appointment_date')
-        customer.appointment_time = data.get('appointment_time')
-        customer.appointment_end_time = data.get('appointment_end_time')
-        customer.status = data.get('status')
-        customer.save()
-        return JsonResponse({'success': True})
-    return JsonResponse({'success': False})
+            # Update the customer with new data
+            customer.person_name = data.get('person_name')
+            customer.phone_number = data.get('phone_number')
+            customer.age = data.get('age')
+            customer.appointment_date = data.get('appointment_date')
+            customer.appointment_time = data.get('appointment_time')
+            customer.appointment_end_time = data.get('appointment_end_time')
+            customer.status = data.get('status')
+            customer.save()
+            return JsonResponse({'success': True})
+        return JsonResponse({'success': False})
+    else:
+        return redirect('login')
 
 
 # function for delete data
 def delete_customer(request, user_id: int):
-    if request.method == 'POST':
-        customer = get_object_or_404(Customer, user_id=user_id)
-        customer.delete()
-        return JsonResponse({'success': True})
-    return JsonResponse({'success': False})
+    if "username" in request.session:
+        if request.method == 'POST':
+            customer = get_object_or_404(Customer, user_id=user_id)
+            customer.delete()
+            return JsonResponse({'success': True})
+        return JsonResponse({'success': False})
+    else:
+        return redirect('login')
